--- apps/codingcatdev/src/routes/(content-single)/course/BookMark.svelte
@@ -1,6 +1,0 @@
-<script lang="ts">
-	import { Icon } from '@steeze-ui/svelte-icon';
-	import { Bookmark } from '@steeze-ui/heroicons';
-</script>
-
-<Icon src={Bookmark} theme="solid" />
