+++ apps/codingcatdev/src/routes/(content-single)/course/LessonCards.svelte
@@ -2,7 +2,7 @@
 	import Image from '$lib/components/content/Image.svelte';
 
 	import type { LayoutData } from './$types';
-	import ProMark from './ProMark.svelte';
+	import ProCourseMark from './ProCourseMark.svelte';
 	export let data: LayoutData;
 </script>
 
@@ -39,7 +39,7 @@
 						<div class="p-4">
 							<header class="relative py-6 text-2xl font-bold capitalize card-header">
 								<div class="absolute top-0 right-0">
-									<ProMark locked={l?.locked} lesson={l} {data} />
+									<ProCourseMark locked={l?.locked} lesson={l} {data} />
 								</div>
 								<div>
 									{l.title}
