+++ apps/codingcatdev/src/routes/(content-single)/course/ProCourseMark.svelte
@@ -2,8 +2,8 @@
 	import Locked from './Locked.svelte';
 	import type { LayoutData } from './$types';
 	import { ContentType, type Lesson } from '$lib/types';
-	import ProComplete from './ProComplete.svelte';
-	import ProBookmark from './ProBookmark.svelte';
+	import ProCourseComplete from './ProCourseComplete.svelte';
+	import ProCourseBookmark from './ProCourseBookmark.svelte';
 	import { auth } from '$lib/client/firebase';
 	import { userStore } from 'sveltefire';
 
@@ -17,8 +17,8 @@
 
 {#if data?.user?.stripeRole && $user?.uid}
 	<div class="flex w-12 gap-1">
-		<ProComplete {data} {lesson} />
-		<ProBookmark {data} {lesson} />
+		<ProCourseComplete {data} {lesson} />
+		<ProCourseBookmark {data} {lesson} />
 	</div>
 {:else if data.content?.type === ContentType.course || data.content?.type === ContentType.lesson}
 	<div class="flex w-6 gap-1">
